+++ fj-doc-playground-quarkus/src/main/java/org/fugerit/java/doc/playground/init/ProjectFacade.java
@@ -1,0 +1,4 @@
+package org.fugerit.java.doc.playground.init;
+
+public class ProjectFacade {
+}