+++ fj-doc-playground-quarkus/src/main/java/org/fugerit/java/doc/playground/init/ProjectInitOutput.java
@@ -1,0 +1,14 @@
+package org.fugerit.java.doc.playground.init;
+
+import lombok.Getter;
+import lombok.Setter;
+
+public class ProjectInitOutput {
+
+    @Getter @Setter
+    private String message;
+
+    @Getter @Setter
+    private String content;
+
+}