+++ fj-doc-playground-quarkus/src/main/react/src/Playground.jsx
@@ -13,6 +13,7 @@ import DocXmlEditor from './playground/DocXmlEditor';
 import DocConversion from './playground/DocConversion';
 import DevValTestForm from './playground/DocValTestForm';
 import DocConfigConvert from './playground/DocConfigConvert';
+import DocProjectInit from './playground/DocProjectInit';
 import { Dialog, DialogTitle, Button, Grid, MenuItem, Select } from "@mui/material";
 
 const homepage = '/fj-doc-playground/home';
@@ -71,7 +72,8 @@ class Playground extends Component {
 					  </Grid>
 					  <Grid item xs={4}>
 							<Select id='toolSelect' defaultValue='def'>
-								<MenuItem id='docValidator' selected={true} value='def'><Grid item xs={4}><Link to={homepage + "/doc_fun/doc_type_validator"}><Button color="primary">Doc Type Validator</Button></Link></Grid></MenuItem>
+								<MenuItem id='docProjectInit' selected={true} value='def'><Grid item xs={4}><Link to={homepage + "/doc_fun/doc_project_init"}><Button color="primary">Doc project init</Button></Link></Grid></MenuItem>
+								<MenuItem id='docValidator' selected={true} value='docValidator'><Grid item xs={4}><Link to={homepage + "/doc_fun/doc_type_validator"}><Button color="primary">Doc Type Validator</Button></Link></Grid></MenuItem>
 								<MenuItem id='docConfigConvert' selected={true} value='xmlToXml'><Grid item xs={4}><Link to={homepage + "/doc_fun/doc_config_convert"}><Button color="primary">Doc config convert</Button></Link></Grid></MenuItem>
 							</Select>
 					  </Grid>
@@ -89,6 +91,7 @@ class Playground extends Component {
 						<Route path={homepage + "/doc_fun/doc_conversion_y2y"} element={<DocConversion handleOpenDialog={this.handleOpenDialog} key='y2y' from='YAML' to='YAML' />} />
 						<Route path={homepage + "/doc_fun/doc_type_validator"} element={<DevValTestForm handleOpenDialog={this.handleOpenDialog} />} />
 						<Route path={homepage + "/doc_fun/doc_config_convert"} element={<DocConfigConvert handleOpenDialog={this.handleOpenDialog} />} />
+						<Route path={homepage + "/doc_fun/doc_project_init"} element={<DocProjectInit handleOpenDialog={this.handleOpenDialog} />} />
 						<Route path="*" element={<Home />} />
 					</Routes>